clean up code, separate into files
