+++ site/components/GraphicsDisplay.tsx
@@ -1,7 +1,9 @@
 import React, { useState } from "react"
 import type { GraphicsObject, Point, Line, Rect, Circle } from "../../lib/types"
+import { GraphicObjectsTable } from "./GraphicObjectsTable"
+import { getTableItemsFromGraphicsObjects } from "../utils/getTableItemsFromGraphicsObjects"
 
-interface GraphicsDisplayProps {
+export interface GraphicsDisplayProps {
   graphics: Array<{
     title: string
     svg: string
@@ -9,7 +11,7 @@ interface GraphicsDisplayProps {
   }>
 }
 
-interface TooltipInfo {
+export interface TooltipInfo {
   x: number
   y: number
   items: Array<{
@@ -17,97 +19,6 @@ interface TooltipInfo {
     label?: string
     position: string
   }>
-}
-
-interface TableObject {
-  type: string
-  id: string
-  properties: Record<string, any>
-}
-
-function flattenObject(obj: any, prefix = ""): Record<string, any> {
-  const result: Record<string, any> = {}
-
-  for (const key of Object.keys(obj)) {
-    const propName = prefix ? `${prefix}.${key}` : key
-    if (typeof obj[key] === "object" && obj[key] !== null) {
-      Object.assign(result, flattenObject(obj[key], propName))
-    } else {
-      result[propName] = obj[key]
-    }
-  }
-
-  return result
-}
-
-function ObjectTable({
-  objects,
-  onHover,
-  svgIndex,
-}: {
-  objects: TableObject[]
-  onHover: (id: string) => void
-  svgIndex: number
-}) {
-  if (objects.length === 0) return null
-
-  const allProperties = Array.from(
-    new Set(objects.flatMap((obj) => Object.keys(obj.properties))),
-  ).sort((a, b) => {
-    if (a === "label") return -1
-    if (b === "label") return 1
-    return a.localeCompare(b)
-  })
-
-  return (
-    <div className="ring-1 ring-gray-200 rounded-lg overflow-hidden">
-      <div className="overflow-x-auto">
-        <table className="min-w-full divide-y divide-gray-200">
-          <thead className="bg-gray-50">
-            <tr>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider whitespace-nowrap">
-                Type
-              </th>
-              {allProperties.map((prop) => (
-                <th
-                  key={prop}
-                  className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider whitespace-nowrap"
-                >
-                  {prop}
-                </th>
-              ))}
-            </tr>
-          </thead>
-          <tbody className="bg-white divide-y divide-gray-200">
-            {objects.map((obj, idx) => (
-              <tr
-                key={`${obj.type}-${idx}`}
-                className="hover:bg-gray-50 cursor-pointer"
-                onMouseEnter={() => {
-                  return onHover(`${obj.type}-${svgIndex}-${idx}`)
-                }}
-                onMouseLeave={() => onHover("")}
-              >
-                <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                  {obj.type}
-                </td>
-                {allProperties.map((prop) => (
-                  <td
-                    key={prop}
-                    className="px-6 py-4 whitespace-nowrap text-sm text-gray-500"
-                  >
-                    {obj.properties[prop] !== undefined
-                      ? String(obj.properties[prop])
-                      : "-"}
-                  </td>
-                ))}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  )
 }
 
 export function GraphicsDisplay({ graphics }: GraphicsDisplayProps) {
@@ -162,64 +73,10 @@ export function GraphicsDisplay({ graphics }: GraphicsDisplayProps) {
     setTooltip(null)
   }
 
-  const processGraphicsObjects = (
-    graphicsObject?: GraphicsObject,
-  ): TableObject[] => {
-    if (!graphicsObject) return []
-
-    const objects: TableObject[] = []
-
-    if (graphicsObject.points) {
-      graphicsObject.points.forEach((point, idx) => {
-        objects.push({
-          type: "point",
-          id: `point-${idx}`,
-          properties: flattenObject(point),
-        })
-      })
-    }
-
-    if (graphicsObject.lines) {
-      graphicsObject.lines.forEach((line, idx) => {
-        objects.push({
-          type: "line",
-          id: `line-${idx}`,
-          properties: {
-            points: `[${line.points.map((p) => `(${p.x},${p.y})`).join(", ")}]`,
-            strokeColor: line.strokeColor || "black",
-            strokeWidth: line.strokeWidth || 1,
-          },
-        })
-      })
-    }
-
-    if (graphicsObject.rects) {
-      graphicsObject.rects.forEach((rect, idx) => {
-        objects.push({
-          type: "rect",
-          id: `rect-${idx}`,
-          properties: flattenObject(rect),
-        })
-      })
-    }
-
-    if (graphicsObject.circles) {
-      graphicsObject.circles.forEach((circle, idx) => {
-        objects.push({
-          type: "circle",
-          id: `circle-${idx}`,
-          properties: flattenObject(circle),
-        })
-      })
-    }
-
-    return objects
-  }
-
   return (
     <div className="space-y-8">
       {graphics.map(({ title, svg, graphicsObject }, index) => {
-        const tableObjects = processGraphicsObjects(graphicsObject)
+        const tableObjects = getTableItemsFromGraphicsObjects(graphicsObject)
 
         return (
           <div key={index} className="space-y-4">
@@ -279,7 +136,7 @@ export function GraphicsDisplay({ graphics }: GraphicsDisplayProps) {
                 )}
               </div>
               <div className="overflow-auto max-h-[640px] border rounded-lg">
-                <ObjectTable
+                <GraphicObjectsTable
                   objects={tableObjects}
                   onHover={setHighlightedId}
                   svgIndex={index}
